pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "process-meteonetwork-retriever"
version = "1.0.38"
description = "Scripts to get data from MeteoNetwork API"
readme = ""
authors = [
  { name="Tommaso Redaelli", email="tommaso.redaelli@gecosistema.com" }
]
license = { text = "MIT" }
keywords = []
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "python-dotenv",
  "requests",
  "filelock",
  "click",
  "boto3",
  "xarray",
  "rioxarray",
  "netcdf4",
  "geopandas"
]

[project.optional-dependencies]
py311gdal = [
  "gdal @ https://github.com/cgohlke/geospatial-wheels/releases/download/v2025.1.20/GDAL-3.10.1-cp311-cp311-win_amd64.whl",
]
pygeoapi = [
  "numpy",
  "numba",
  "gdal2numpy",
  "pygeoapi",
]

[project.urls]
Homepage = "https://github.com/SaferPlaces2023/process-meteonetwork-retriever"

[project.scripts]
meteonetwork-retriever = "process_meteonetwork_retriever.main:cli_run_meteonetwork_retriever"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
